Support for stereoscopic video.
Apparently, Oculus/Samsung developers forgot to port stereoscopic video
support to release 4.0.0 (the latest stable one in the framework).
To avoid including a reference to a local version of the SDK,
I have created a "fixed" version of the VideoSceneObject which
calls the right shaders. Details on the fix here:
https://github.com/Samsung/GearVRf/pull/1672